requirements.txt
asynch>=0.2.5
blinker>=1.9.0
certifi>=2025.1.31
charset-normalizer>=3.4.1
ciso8601>=2.3.2
click>=8.1.8
clickhouse-driver>=0.2.9
clickhouse-sqlalchemy>=0.3.2
Flask>=3.1.0
idna>=3.10
itsdangerous>=2.2.0
Jinja2>=3.1.6
leb128>=1.0.8
lz4>=4.4.3
MarkupSafe>=3.0.2
psycopg2-binary>=2.9.10
python-dotenv>=1.1.0
pytz>=2025.2
requests>=2.32.3
SQLAlchemy>=2.0.40
typing_extensions>=4.13.0
tzlocal>=5.3.1
urllib3>=2.3.0
Werkzeug>=3.1.3
zstd>=1.5.6.6
